Move game tests to dedicated neoforge-gametest subproject
Extract game test code (StorageGameTests, GameTestStructureProvider, S3DataGeneration)
from neoforge module into a new neoforge-gametest Gradle subproject. This separates
test code from production code and gives game tests an independent build lifecycle.

Closes #31
