+++ .github/workflows/game-tests.yml
@@ -26,7 +26,7 @@ jobs:
         run: chmod +x gradlew
 
       - name: Generate data (structures for game tests)
-        run: ./gradlew :neoforge:runData
+        run: ./gradlew :neoforge-gametest:runData
 
       - name: Run game tests
-        run: ./gradlew :neoforge:runGameTestServer
+        run: ./gradlew :neoforge-gametest:runGameTestServer
