+++ neoforge/build.gradle
@@ -16,20 +16,6 @@ neoForge {
         server {
             server()
         }
-        gameTestServer {
-            type = "gameTestServer"
-            systemProperty 'neoforge.enabledGameTestNamespaces', 's3'
-        }
-        data {
-            data()
-            programArguments.addAll(
-                '--mod', 's3',
-                '--all',
-                '--output', file('src/generated/resources/').getAbsolutePath(),
-                '--existing', file('src/main/resources/').getAbsolutePath(),
-                '--existing', rootProject.file('core/src/main/resources/').getAbsolutePath()
-            )
-        }
     }
 
     mods {
@@ -44,8 +30,6 @@ neoForge {
 jar {
     from project(':core').sourceSets.main.output
 }
-
-sourceSets.main.resources.srcDir 'src/generated/resources'
 
 processResources {
     def replaceProperties = [mod_version: mod_version]
